feat: stylex direct
