+++ frontend/apps/web/components/server/Spinner/Spinner.tsx
@@ -1,6 +1,4 @@
-import { colorTokens, animations } from '@/styles/index.stylex';
 import stylex from '@stylexjs/stylex';
-import React from 'react';
 
 interface SpinnerProps {
   size?: 'medium' | 'large';
@@ -23,6 +21,18 @@ const Spinner = ({ size = 'medium' }: SpinnerProps) => {
 
 export default Spinner;
 
+const rotateAnimation = stylex.keyframes({
+  '0%': {
+    transform: 'rotate(0deg) translateX(20px) rotate(0deg)',
+  },
+  '50%': {
+    transform: 'rotate(180deg) translateX(20px) rotate(-180deg)',
+  },
+  '100%': {
+    transform: 'rotate(360deg) translateX(20px) rotate(-360deg)',
+  },
+});
+
 const styles = stylex.create({
   spinner: (size: 'medium' | 'large') => ({
     display: 'inline-flex',
@@ -36,9 +46,9 @@ const styles = stylex.create({
     position: 'absolute',
     width: `${sizeMap[size].circle}px`,
     height: `${sizeMap[size].circle}px`,
-    backgroundColor: colorTokens.primary,
+    backgroundColor: '#F5B343',
     borderRadius: '50%',
-    animationName: animations.rotateAnimation,
+    animationName: rotateAnimation,
     animationDuration: '5s',
     animationTimingFunction: 'linear',
     animationIterationCount: 'infinite',
@@ -48,7 +58,7 @@ const styles = stylex.create({
   centerCircle: (size: 'medium' | 'large') => ({
     width: `${sizeMap[size].centerCircle}px`,
     height: `${sizeMap[size].centerCircle}px`,
-    backgroundColor: colorTokens.primary,
+    backgroundColor: '#F5B343',
     borderRadius: '50%',
   }),
 });
