--- frontend/apps/web/styles/index.stylex.ts
@@ -1,2 +1,0 @@
-export * from './animations.stylex';
-export * from './colorTokens.stylex';
